requirements.txt
python-dotenv==0.20.0
gunicorn==20.1.0
marshmallow==3.17.1

Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Bcrypt==1.0.1

pytest==7.1.2
